requirements.txt
Django==2.1
django-appconf==1.0.2
django-imagekit==4.0.2
pilkit==2.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
six==1.11.0
